Move community share your stuff into new file
