+++ src/components/pages/home/index.js
@@ -1,5 +1,6 @@
 import AcademiaFeature from './academia-feature';
+import CommunityFeature from './community-feature';
 import EventsFeature from './events-feature';
 import TwitchFeature from './twitch-feature';
 
-export { AcademiaFeature, EventsFeature, TwitchFeature };
+export { AcademiaFeature, CommunityFeature, EventsFeature, TwitchFeature };
